Refactor DataTable to be configurable and reusable
