+++ src/components/DataTable.config.tsx
@@ -1,0 +1,261 @@
+import React from 'react';
+import { TableEntry, POS_LABELS } from '@/lib/types';
+import { ColumnConfig } from './ColumnVisibilityPanel';
+
+// Column width state interface
+export interface ColumnWidthState {
+  [columnKey: string]: number;
+}
+
+// Default columns for verbs table
+export const VERBS_COLUMNS: ColumnConfig[] = [
+  { key: 'id', label: 'ID', visible: true, sortable: true },
+  { key: 'legacy_id', label: 'Legacy ID', visible: false, sortable: true },
+  { key: 'lemmas', label: 'Lemmas', visible: true, sortable: true },
+  { key: 'gloss', label: 'Definition', visible: true, sortable: true },
+  { key: 'pos', label: 'Part of Speech', visible: false, sortable: true },
+  { key: 'lexfile', label: 'Lexfile', visible: true, sortable: true },
+  { key: 'isMwe', label: 'Multi-word Expression', visible: false, sortable: true },
+  { key: 'transitive', label: 'Transitive', visible: false, sortable: true },
+  { key: 'flagged', label: 'Flagged', visible: false, sortable: true },
+  { key: 'forbidden', label: 'Forbidden', visible: false, sortable: true },
+  { key: 'particles', label: 'Particles', visible: false, sortable: false },
+  { key: 'frames', label: 'Frames', visible: false, sortable: false },
+  { key: 'examples', label: 'Examples', visible: false, sortable: false },
+  { key: 'parentsCount', label: 'Parents', visible: true, sortable: true },
+  { key: 'childrenCount', label: 'Children', visible: true, sortable: true },
+  { key: 'createdAt', label: 'Created', visible: false, sortable: true },
+  { key: 'updatedAt', label: 'Updated', visible: false, sortable: true },
+];
+
+// Default column widths for verbs table
+export const VERBS_COLUMN_WIDTHS: ColumnWidthState = {
+  id: 120,
+  legacy_id: 150,
+  lemmas: 150,
+  gloss: 300,
+  pos: 120,
+  lexfile: 120,
+  isMwe: 100,
+  transitive: 100,
+  flagged: 100,
+  forbidden: 100,
+  particles: 120,
+  frames: 100,
+  examples: 250,
+  parentsCount: 150,
+  childrenCount: 150,
+  createdAt: 100,
+  updatedAt: 100,
+};
+
+// Helper functions for rendering
+const truncateText = (text: string, maxLength: number) => {
+  if (text.length <= maxLength) return text;
+  return text.substring(0, maxLength) + '...';
+};
+
+const formatDate = (date: Date) => {
+  return new Date(date).toLocaleDateString();
+};
+
+// Cell renderer for verbs table
+export const renderVerbsCell = (
+  entry: TableEntry,
+  columnKey: string,
+  relationsData?: { parents: string[]; children: string[] }
+): React.ReactNode => {
+  switch (columnKey) {
+    case 'lemmas':
+      // Concatenate src_lemmas and lemmas for display
+      const allLemmas = [...(entry.src_lemmas || []), ...(entry.lemmas || [])];
+      return (
+        <div className="flex flex-wrap gap-1">
+          {allLemmas.slice(0, 3).map((lemma, idx) => (
+            <span 
+              key={idx}
+              className="inline-block px-2 py-1 text-xs bg-blue-100 text-blue-800 rounded"
+            >
+              {lemma}
+            </span>
+          ))}
+          {allLemmas.length > 3 && (
+            <span className="inline-block px-2 py-1 text-xs bg-gray-100 text-gray-600 rounded">
+              +{allLemmas.length - 3}
+            </span>
+          )}
+        </div>
+      );
+    case 'gloss':
+      return (
+        <div className="text-sm text-gray-900" title={entry.gloss}>
+          {truncateText(entry.gloss, 150)}
+        </div>
+      );
+    case 'pos':
+      return (
+        <span className="inline-block px-2 py-1 text-xs bg-green-100 text-green-800 rounded font-medium">
+          {POS_LABELS[entry.pos as keyof typeof POS_LABELS] || entry.pos}
+        </span>
+      );
+    case 'lexfile':
+      return <span className="text-xs text-gray-500">{entry.lexfile.replace(/^verb\./, '')}</span>;
+    case 'isMwe':
+      return (
+        <span className={`inline-block px-2 py-1 text-xs rounded font-medium ${
+          entry.isMwe 
+            ? 'bg-purple-100 text-purple-800' 
+            : 'bg-gray-100 text-gray-600'
+        }`}>
+          {entry.isMwe ? 'Yes' : 'No'}
+        </span>
+      );
+    case 'transitive':
+      if (entry.transitive === null || entry.transitive === undefined) {
+        return <span className="text-gray-400 text-sm">N/A</span>;
+      }
+      return (
+        <span className={`inline-block px-2 py-1 text-xs rounded font-medium ${
+          entry.transitive 
+            ? 'bg-blue-100 text-blue-800' 
+            : 'bg-gray-100 text-gray-600'
+        }`}>
+          {entry.transitive ? 'Yes' : 'No'}
+        </span>
+      );
+    case 'flagged':
+      if (entry.flagged === null || entry.flagged === undefined) {
+        return <span className="text-gray-400 text-sm">N/A</span>;
+      }
+      return (
+        <span className={`inline-block px-2 py-1 text-xs rounded font-medium ${
+          entry.flagged 
+            ? 'bg-orange-100 text-orange-800' 
+            : 'bg-gray-100 text-gray-600'
+        }`}>
+          {entry.flagged ? 'Yes' : 'No'}
+        </span>
+      );
+    case 'forbidden':
+      if (entry.forbidden === null || entry.forbidden === undefined) {
+        return <span className="text-gray-400 text-sm">N/A</span>;
+      }
+      return (
+        <span className={`inline-block px-2 py-1 text-xs rounded font-medium ${
+          entry.forbidden 
+            ? 'bg-red-100 text-red-800' 
+            : 'bg-gray-100 text-gray-600'
+        }`}>
+          {entry.forbidden ? 'Yes' : 'No'}
+        </span>
+      );
+    case 'particles':
+      if (!entry.particles || entry.particles.length === 0) {
+        return <span className="text-gray-400 text-sm">None</span>;
+      }
+      return (
+        <div className="flex flex-wrap gap-1">
+          {entry.particles.slice(0, 2).map((particle, idx) => (
+            <span 
+              key={idx}
+              className="inline-block px-2 py-1 text-xs bg-orange-100 text-orange-800 rounded"
+            >
+              {particle}
+            </span>
+          ))}
+          {entry.particles.length > 2 && (
+            <span className="inline-block px-2 py-1 text-xs bg-gray-100 text-gray-600 rounded">
+              +{entry.particles.length - 2}
+            </span>
+          )}
+        </div>
+      );
+    case 'frames':
+      if (!entry.frames || entry.frames.length === 0) {
+        return <span className="text-gray-400 text-sm">None</span>;
+      }
+      return (
+        <div className="text-xs text-gray-600" title={entry.frames.join(', ')}>
+          {entry.frames.length} frame{entry.frames.length !== 1 ? 's' : ''}
+        </div>
+      );
+    case 'examples':
+      if (!entry.examples || entry.examples.length === 0) {
+        return <span className="text-gray-400 text-sm">None</span>;
+      }
+      return (
+        <div className="space-y-1 text-xs text-gray-700 max-w-md">
+          {entry.examples.map((example, idx) => (
+            <div key={idx} className="leading-relaxed">
+              <span className="text-gray-400 mr-1">{idx + 1}.</span>
+              {example}
+            </div>
+          ))}
+        </div>
+      );
+    case 'parentsCount':
+      if (!relationsData?.parents || relationsData.parents.length === 0) {
+        return <span className="text-gray-400 text-sm">None</span>;
+      }
+      return (
+        <div className="space-y-1 text-xs text-gray-700 max-w-sm">
+          {relationsData.parents.map((parentId, idx) => (
+            <div key={idx} className="font-mono text-blue-600">
+              {parentId}
+            </div>
+          ))}
+        </div>
+      );
+    case 'childrenCount':
+      if (!relationsData?.children || relationsData.children.length === 0) {
+        return <span className="text-gray-400 text-sm">None</span>;
+      }
+      return (
+        <div className="space-y-1 text-xs text-gray-700 max-w-sm">
+          {relationsData.children.map((childId, idx) => (
+            <div key={idx} className="font-mono text-green-600">
+              {childId}
+            </div>
+          ))}
+        </div>
+      );
+    case 'id':
+      return <span className="text-sm font-mono text-blue-600">{entry.id}</span>;
+    case 'legacy_id':
+      return <span className="text-sm font-mono text-gray-600">{entry.legacy_id}</span>;
+    case 'createdAt':
+      return <span className="text-xs text-gray-500">{formatDate(entry.createdAt)}</span>;
+    case 'updatedAt':
+      return <span className="text-xs text-gray-500">{formatDate(entry.updatedAt)}</span>;
+    default:
+      return <span className="text-sm text-gray-900">{String((entry as unknown as Record<string, unknown>)[columnKey] || '')}</span>;
+  }
+};
+
+// Function to fetch relations for an entry (verb-specific logic)
+export const fetchVerbRelations = async (entryId: string): Promise<{ parents: string[]; children: string[] }> => {
+  try {
+    const response = await fetch(`/api/entries/${entryId}/relations`);
+    if (!response.ok) {
+      throw new Error('Failed to fetch relations');
+    }
+    const data = await response.json();
+    
+    // Extract parent IDs (hypernyms - more general concepts this entry points to)
+    const parents = data.sourceRelations
+      .filter((rel: { type: string }) => rel.type === 'hypernym')
+      .map((rel: { target?: { id: string } }) => rel.target?.id)
+      .filter(Boolean);
+    
+    // Extract children IDs (hyponyms - more specific concepts that point to this entry)
+    const children = data.targetRelations
+      .filter((rel: { type: string }) => rel.type === 'hypernym')
+      .map((rel: { source?: { id: string } }) => rel.source?.id)
+      .filter(Boolean);
+    
+    return { parents, children };
+  } catch (error) {
+    console.error('Error fetching relations:', error);
+    return { parents: [], children: [] };
+  }
+};
